refactor: move router folder to correct location (#22)
